[Common] - Fix - inuring changes(discussed with Soufiane)

+++ RR-Common/src/main/java/com/scor/rr/exceptions/inuring/InputPLTNotFoundException.java
@@ -7,7 +7,7 @@ import com.scor.rr.exceptions.RRException;
  * Created by u004602 on 11/09/2019.
  */
 public class InputPLTNotFoundException extends RRException {
-    public InputPLTNotFoundException(int pltId) {
+    public InputPLTNotFoundException(long pltId) {
         super(ExceptionCodename.PLT_NOT_FOUND, "Input PLT id " + pltId + " not found");
     }
 }
